requirements.txt
pandas~=1.4.3
requests~=2.28.1
python-dotenv~=0.20.0